Add files via upload
This is a file mainly focusing on data cleaning. It contains real world data with typo in data as it is acquired from a survey. This is not the finest work due to time constraints and there is a large scope for improvement. Will post the updated and better version soon.